+++ app/services/waste_carriers_engine/govpay.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
 require_relative "./govpay/object"
-require_relative "./govpay/objects/error"
-require_relative "./govpay/objects/refund"
-require_relative "./govpay/objects/payment"
+require_relative "./govpay/error"
+require_relative "./govpay/refund"
+require_relative "./govpay/payment"
